app/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
et-xmlfile==1.1.0
Flask==2.3.2
frozenlist==1.3.3
google-api-core==2.11.1
google-api-python-client==2.91.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
oauthlib==3.2.2
openpyxl==3.1.2
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyee==11.0.0
pyparsing==3.1.0
python-dotenv==1.0.0
random-password-generator==2.2.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
slack-sdk==3.21.3
slackclient==2.9.4
slackeventsapi==3.0.1
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==2.3.6
yarl==1.9.2
